move sqlparser test runner to tests dir
