--- src/sqlparser/test_runner/Cargo.toml
@@ -1,42 +1,0 @@
-[package]
-name = "risingwave_sqlparser_test_runner"
-version = "0.1.0"
-edition = "2021"
-
-[package.metadata.cargo-machete]
-ignored = ["workspace-hack"]
-
-[package.metadata.cargo-udeps.ignore]
-normal = ["workspace-hack"]
-
-[dependencies]
-anyhow = "1"
-console = "0.15"
-futures = { version = "0.3", default-features = false, features = ["alloc"] }
-risingwave_sqlparser = { workspace = true }
-serde = { version = "1", features = ["derive"] }
-serde_with = "3"
-serde_yaml = "0.9"
-tokio = { version = "0.2", package = "madsim-tokio", features = [
-    "rt",
-    "rt-multi-thread",
-    "sync",
-    "macros",
-    "time",
-    "signal",
-    "fs",
-] }
-walkdir = "2"
-
-[[bin]]
-name = "parser-test-apply"
-path = "src/bin/apply.rs"
-
-[target.'cfg(not(madsim))'.dependencies]
-workspace-hack = { path = "../../workspace-hack" }
-
-[build-dependencies]
-walkdir = "2"
-
-[lints]
-workspace = true
