--- src/sqlparser/test_runner/sqlparser_test.toml
@@ -1,56 +1,0 @@
-[tasks.update-parser-test]
-description = "Update parser test data"
-private = true
-script = '''
-#!/usr/bin/env bash
-set -e
-cargo run --bin parser-test-apply
-'''
-
-[tasks.apply-parser-test]
-private = true
-description = "Generate parser test data"
-dependencies = [
-  "update-parser-test"
-]
-script = '''
-#!/usr/bin/env bash
-set -e
-cd src/sqlparser/tests/testdata/
-
-for f in *.apply.yaml
-do
-  diff "$f" "$(basename "$f" .apply.yaml).yaml" || true
-done
-
-echo "If you want to apply the parser test data, run: $(tput setaf 2)./risedev do-apply-parser-test$(tput sgr 0)"
-'''
-category = "RiseDev - Test"
-
-[tasks.do-apply-parser-test]
-description = "Apply parser test data"
-dependencies = [
-  "update-parser-test"
-]
-script = '''
-#!/usr/bin/env bash
-set -e
-cd src/sqlparser/tests/testdata/
-
-for f in *.apply.yaml
-do
-  SOURCE="$(basename $f .apply.yaml).yaml"
-  if [ -f "$SOURCE" ]; then
-    cat <<EOF > temp.apply.yaml
-# This file is automatically generated. See \`src/sqlparser/test_runner/src/bin/apply.rs\` for more information.
-EOF
-    cat "$f" >> temp.apply.yaml
-    mv temp.apply.yaml "$SOURCE"
-  fi
-done
-
-rm *.apply.yaml
-
-echo "$(tput setaf 2)Diff applied!$(tput sgr 0)"
-'''
-category = "RiseDev - Test"
