+++ src/sqlparser/tests/testdata/alter.yaml
@@ -1,4 +1,4 @@
-# This file is automatically generated. See `src/sqlparser/test_runner/src/bin/apply.rs` for more information.
+# This file is automatically generated by `src/sqlparser/tests/parser_test.rs`.
 - input: ALTER USER user WITH SUPERUSER CREATEDB PASSWORD 'password'
   formatted_sql: ALTER USER user WITH SUPERUSER CREATEDB PASSWORD 'password'
 - input: ALTER USER user RENAME TO another
