+++ src/sqlparser/tests/testdata/array.yaml
@@ -1,4 +1,4 @@
-# This file is automatically generated. See `src/sqlparser/test_runner/src/bin/apply.rs` for more information.
+# This file is automatically generated by `src/sqlparser/tests/parser_test.rs`.
 - input: CREATE TABLE t(a int[]);
   formatted_sql: CREATE TABLE t (a INT[])
 - input: CREATE TABLE t(a int[][]);
