+++ src/sqlparser/tests/testdata/privilege.yaml
@@ -1,4 +1,4 @@
-# This file is automatically generated. See `src/sqlparser/test_runner/src/bin/apply.rs` for more information.
+# This file is automatically generated by `src/sqlparser/tests/parser_test.rs`.
 - input: GRANT ALL ON DATABASE database TO user1 WITH GRANT OPTION GRANTED BY user
   formatted_sql: GRANT ALL ON DATABASE database TO user1 WITH GRANT OPTION GRANTED BY user
   formatted_ast: 'Grant { privileges: All { with_privileges_keyword: false }, objects: Databases([ObjectName([Ident { value: "database", quote_style: None }])]), grantees: [Ident { value: "user1", quote_style: None }], with_grant_option: true, granted_by: Some(Ident { value: "user", quote_style: None }) }'
