+++ src/sqlparser/tests/testdata/qualified_operator.yaml
@@ -1,4 +1,4 @@
-# This file is automatically generated. See `src/sqlparser/test_runner/src/bin/apply.rs` for more information.
+# This file is automatically generated by `src/sqlparser/tests/parser_test.rs`.
 - input: select 2 * 3 operator(pg_catalog.+) 4;
   formatted_sql: SELECT 2 * 3 OPERATOR(pg_catalog.+) 4
   formatted_ast: 'Query(Query { with: None, body: Select(Select { distinct: All, projection: [UnnamedExpr(BinaryOp { left: BinaryOp { left: Value(Number("2")), op: Multiply, right: Value(Number("3")) }, op: PGQualified(QualifiedOperator { schema: Some(Ident { value: "pg_catalog", quote_style: None }), name: "+" }), right: Value(Number("4")) })], from: [], lateral_views: [], selection: None, group_by: [], having: None }), order_by: [], limit: None, offset: None, fetch: None })'
