+++ src/sqlparser/tests/testdata/set.yaml
@@ -1,4 +1,4 @@
-# This file is automatically generated. See `src/sqlparser/test_runner/src/bin/apply.rs` for more information.
+# This file is automatically generated by `src/sqlparser/tests/parser_test.rs`.
 - input: SET TIME ZONE LOCAL
   formatted_sql: SET TIME ZONE LOCAL
 - input: SET TIME ZONE DEFAULT
