+++ src/sqlparser/tests/testdata/show.yaml
@@ -1,4 +1,4 @@
-# This file is automatically generated. See `src/sqlparser/test_runner/src/bin/apply.rs` for more information.
+# This file is automatically generated by `src/sqlparser/tests/parser_test.rs`.
 - input: SHOW DATABASES
   formatted_sql: SHOW DATABASES
   formatted_ast: 'ShowObjects { object: Database, filter: None }'
