+++ src/sqlparser/tests/testdata/struct.yaml
@@ -1,4 +1,4 @@
-# This file is automatically generated. See `src/sqlparser/test_runner/src/bin/apply.rs` for more information.
+# This file is automatically generated by `src/sqlparser/tests/parser_test.rs`.
 - input: SELECT (1 * 2, 1.0)::foo;
   formatted_sql: SELECT CAST(ROW(1 * 2, 1.0) AS foo)
 - input: SELECT ROW(1 * 2, 1.0)::foo;
